Created a package to hold all user interaction classes. Completed the initial code for the user class.
